requirements.txt
fastapi 
uvicorn 
requests 
scikit-learn 
numpy 
python-dotenv
huggingface_hub